Use video URL with largest size
For some videos (looks like it's about .mov files), Flickr provides
a Video Original URL which does not work. So, as a workaround, I
decided to always pickup the source URL which has the largets quality.
This is not necessarily the Original Video URL.
For example, I have this situation:

<size label="Site MP4" width="640" height="481" ...
<size label="Video Original" width="640" height="480" ...

I don't know why Site MP4 has one extra unit in height but, it allows
the program to decide to use Site MP4 instead of Video Original.
This is good (compromise) since the Video Original URL is not working anyway.

Maybe I will also get something out of this forum:
https://www.flickr.com/groups/51035612836@N01/discuss/72157671986445591/72157667700617229/